Formatting: A way to stop padding around binary operators
The common way to write mathematical operators are to group them by
their priority. The reason is to increase the readability of the
numerical operators. This is not only seen when humans write equations
with pen and paper but also recommended by PEP8 and OpenFOAM coding
style.

This commit introduces an option to stop padding multiplication and
division operators. This commit can be easily extended for other
operators, i.e., modulo operator